// frameworks/bridge/arkts_frontend/koala_mirror/arkoala-arkts/arkui/src/generated/arkui-external.ts
export class SymbolEffect implements MaterializedBase {
    peer?: Finalizable | undefined = undefined
    public getPeer(): Finalizable | undefined {
        return this.peer
    }
    static ctor_symboleffect(): KPointer {
        const retval  = ArkUIGeneratedNativeModule._SymbolEffect_ctor()
        return retval
    }
    constructor() {
        const ctorPtr : KPointer = SymbolEffect.ctor_symboleffect()
        this.peer = new Finalizable(ctorPtr, SymbolEffect.getFinalizer())
    }
    static getFinalizer(): KPointer {
        return ArkUIGeneratedNativeModule._SymbolEffect_getFinalizer()
    }
    public dummyForAccessorGenerate(): void {
        this.dummyForAccessorGenerate_serialize()
        return
    }
    private dummyForAccessorGenerate_serialize(): void {
        ArkUIGeneratedNativeModule._SymbolEffect_dummyForAccessorGenerate(this.peer!.ptr)
    }
}
// ...
export enum EffectDirection {
    DOWN = 0,
    UP = 1
}
export enum EffectScope {
    LAYER = 0,
    WHOLE = 1
}